requirements.txt
pillow==10.2.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
typing_extensions==4.10.0
validators==0.24.0